Add files via upload
The inventory management system displays the features of a selected shoe along with its quantity amount. The display can be sorted out by the different attributes of the shoe. The features range from its colour, shoe size, model name, and model number. For example, if a consumer would like to view all products with the size US 6.5, it will display out all the shoes that have a size of 6.5, showing the necessary information of the shoes. 

The billing system calculates the total amount the consumer would like to purchase. This function will prompt the user with a list of the model names, colour, and amount for the consumer to choose. After choosing the desired product, another list will be shown of the selected model name and colour. The difference is it will also display the shoe size of the chosen item as well as its model number. Shoe size is shown after selecting the products to make the list more user-friendly, as it displays only the necessary information. The amount will then be added to the total bill, and the loop continues until the user would like to end their purchase. The total calculated amount will be displayed to the user. Meanwhile, the quantity of the products that have been brought will decrease by itself, showing the newly updated quantity of the inventory.